# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Flash-Archi Public API
  description: >
    API publique de Flash-Archi (pré-production) — projets, jobs,
    artefacts, versions et webhooks.

    Clés d'accès :
    - **API publique** (`/api/public/v1/**`) : clé Bearer étendue avec scopes
      `public:*`, résolue par tenant (plan : free / pro / enterprise).
    - **API interne** (`/api/flash-archi/projects/**/versions/**`)
      : clé applicative `API_KEY` (environnement), scopes internes.

    Tous les endpoints honorent le rate limiting (headers `X-RateLimit-*`).
  version: 3.3.0
  contact:
    name: Flash-Archi
servers:
- url: /
  description: "Racine \u2014 chemins complets (API publique + interne)"
...
paths:
...
  /api/public/v1/webhooks:
    get:
      tags:
      - Webhooks
      summary: Lister les abonnements webhook (sans secrets)
      operationId: listWebhooks
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      responses:
        '200':
          description: Liste des webhooks du tenant
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookSubscription'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      tags:
      - Webhooks
      summary: "Créer un abonnement webhook (le secret est retourné une seule fois)"
      operationId: createWebhook
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - url
              - events
              properties:
                url:
                  type: string
                  format: uri
                  description: "URL de r\xE9ception du webhook"
                  example: https://hooks.example.com/fa
                events:
                  type: array
                  items:
                    type: string
                    enum:
                    - job.completed
                    - job.failed
                    - project.created
                    - project.updated
                  minItems: 1
                  example:
                  - job.completed
                  - job.failed
      responses:
        '201':
          description: "Webhook cr\xE9\xE9"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookWithSecret'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/public/v1/webhooks/{webhookId}:
    get:
      tags:
      - Webhooks
      summary: "Lire un webhook (secret masqu\xE9)"
      operationId: getWebhook
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      parameters:
      - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Webhook (sans secret)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscription'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      tags:
      - Webhooks
      summary: Modifier un webhook (url, events, active)
      operationId: updateWebhook
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      parameters:
      - $ref: '#/components/parameters/WebhookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum:
                    - job.completed
                    - job.failed
                    - project.created
                    - project.updated
                active:
                  type: boolean
              minProperties: 1
      responses:
        '200':
          description: "Webhook mis \xE0 jour (sans secret)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags:
      - Webhooks
      summary: Supprimer un abonnement webhook
      operationId: deleteWebhook
      security:
      - PublicKeyAuth:
        - public:manage:webhooks
      parameters:
      - $ref: '#/components/parameters/WebhookId'
      responses:
        '204':
          description: "Webhook supprim\xE9"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    PublicKeyAuth:
      type: http
      scheme: bearer
      description: "Cl\xE9 API publique (scope `public:*`) r\xE9solue par tenant.
        \ Le scope requis est indiqu\xE9 par endpoint. Plans : free=100 req/min, pro=500
        \ req/min, enterprise=1000 req/min.\n"
    InternalKeyAuth:
      type: http
      scheme: bearer
      description: "Cl\xE9 applicative interne `API_KEY` (routes versions)."
  parameters:
...
    WebhookId:
      name: webhookId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Identifiant du webhook
...
  schemas:
...
    WebhookSubscription:
      type: object
      description: "Abonnement webhook (le secret n'est JAMAIS expos\xE9)"
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
            - job.completed
            - job.failed
            - project.created
            - project.updated
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WebhookWithSecret:
      allOf:
      - $ref: '#/components/schemas/WebhookSubscription'
      - type: object
        description: "Retourn\xE9 UNIQUEMENT \xE0 la cr\xE9ation"
        required:
        - secret
        properties:
          secret:
            type: string
            description: "Cl\xE9 HMAC-SHA256 \u2014 \xE0 conserver, non r\xE9cup\xE9
              rable ensuite"
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: "Requ\xEAte invalide"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: "Authentification \xE9chou\xE9e ou cl\xE9 manquante"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Scope insuffisant pour cette ressource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: "Ressource introuvable (ou hors du tenant \u2014 RLS, 404 volontaire)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: "Quota d\xE9pass\xE9"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: "Quota de la fen\xEAtre"
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: "Requ\xEAtes restantes dans la fen\xEAtre"
        X-RateLimit-Reset:
          schema:
            type: integer
          description: "Timestamp (ms) de r\xE9initialisation"
        Retry-After:
          schema:
            type: integer
          description: "Secondes avant retry autoris\xE9"